requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.3
cron-descriptor==1.4.3
decorator==5.1.1
Django==4.2.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==23.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.35.2
executing==2.0.1
greenlet==3.0.3
idna==3.6
inflection==0.5.1
ipython==8.22.1
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
